Corrected script syntax and tw=0 in vimrc
The syntax of the _macosx.sh script is corrected so that it can handle spaces
in the directory names better. Also, comments were added and the overall script
was cleaned. In addition, for the Mac OS X version of _vimrc is changed tw=120
to tw=0 to prevent automatic new lines.