Mac: less news urls, vim ruler; T440s: tab size
For the Mac OSX configs, the urls in newsbeuter is reduced to something more
manageable. Also, the vimrc is changed so that it always displays the column
numbers on the bottom right. For T440s config files, the vimperatorrc is
modified so that it doesn't try to change the tab height, which was causing
some graphical issues when creating new tabs.