Special vim settings for cpp (Mac OSX)
Added in filetype specific settings in the vimrc for only cpp code. Mainly, it
turns off expandtab (for code consistency with group), even though expandtab is
the current default I prefer. This setting will be played around for a bit in
Mac OSX before I apply it to all the other machines.