Added in library path for remote GNU/Linux servers
Not sure if this works, but included the library path in the bashrc dotfile so
that it may (potentially?) be able to read a local directory of library files.
If this works, it will be particularly useful in local installation of various
software... If not, there's no harm in having an extra environmental variable.