Change to all newsbeuter feed lists
For Mac OSX and T440s, the newsbeuter feed lists have been updated so that the
local news is separated for visual purposes and the default Google feeds are
removed in favor of the custom ones handled by my own server. This commit is
made after the change to the Python Readability scripts on the server.