Fixed status links on Mac OSX vimperator config
Adjusted the config file of vimperator in Mac OSX so that the status links are
shown properly in the bottom statusbar. Note: The options for showstatuslinks
on Mac OSX are 0 (no link), 1 (status line), and 2 (show in command line). Not
sure if the settings are the same on GNU/Linux since the previously made T440s
config file sets the value to 3...