Set all quit shortcuts to current directory in vifm
In vifm (currently only on the remote server), all quit shortcuts (i.e. ZZ and
ZQ) now leave the user in the current directory rather than in the original.
The option will stay for a few days to see how it affects the workflow, at
which point a decision will be made as to whether to keep the options.