Sanitized git config
Unfortunately, past commits had some personal information in the global
settings revealed. This has been cleaned up so that at least the
master/main version of the code is sanitized. At this point, history is
history so keeping it for the time being. May want to move repo into a
private setting later or re-initialize the repo from scratch...